hackrx_app/requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
PyMuPDF
tqdm
faiss-cpu
numpy
